fix: update 'get' to resolve paths
Prior to this change the dag::get call would not go to the network to
find a CID, but would only check local storage. Now dag::get will
resolve the CID with the network and populate local storage with the
result.

Additionally a list of Ceramic specific bootstrap servers has been
provided.