feat: add /metrics endpoint
This change exposes a Prometheus /metrics endpoint to enable metric
collection via scraping.

Additionally a refactor to the how Beetle is consumed was performed.
It was becoming very time consuming to trace Beetle code paths through
many unneeded abstractions (i.e. unixfs). As such the refactor means
that we now consume a minimal set of dependent crates from Beetle.

This has the advantage that is now very clear how these systems
interact. For example publishing DHT records or when/how IPLD blocks
are added to the local store.

Specifically we now depend on:

* iroh-metrics
* iroh-p2p
* iroh-rpc-client
* iroh-rpc-types
* iroh-store

Previously we were dependent on:

* iroh-api
* iroh-bitswap
* iroh-car
* iroh-embed
* iroh-gateway
* iroh-metrics
* iroh-one
* iroh-p2p
* iroh-resolver
* iroh-rpc-client
* iroh-rpc-types
* iroh-store
* iroh-unixfs
* iroh-util

This change comes with some small duplicated code to setup
the p2p and store services. However this is a small amount
of code with a clear purpose.