refactor: remove mutex locking of Recon
This change introduces a client/server model for communicating with
Recon. The Server runs as a single async task and the clients send
requests to the Server via async channels.

This change in design removes the need to use an Arc<Mutex<Recon>> as
the client can be cheaply cloned and owned by the callers. As a result
this change moves us towards a full async stack top to bottom so we are
not mixing sync and async code.

Additionally rusqlite was replaced with sqlx to make the entire stack
async.