refactor: move beetle locally
This change moves all beetle/iroh dependencies into this repo.
This will enable easier review of bug fixes and allow for publishing
this crate on crates.io without having to also publish our beetle fork
on crates.io. We do not intend for our fork of beetle to be consumed
externally so this is no loss of functionality.