fix: change bitswap dial behavior
Prior to this change bitswap assumed it was the only behaviour asking
the swarm to dial peers, as such when a dial failure occurred it marked
the peer as failed and would not send want-haves to the peer.

However this behavior did not work well with the behavior to auto
discover ceramic peers and dial them. When ceramic peers were discovered
it was possible that they were already connected in which case the
`DialFailedPrecondition` error would be reported causing bitswap to mark
the peer as failed even though we have a valid connection open to the
peer.

This change modifies the bitswap behavior to only modify peer state when
a connection closed event is received instead of dial failures. A peer
is not marked as good until a connection is established so marking a
peer as failed when the connection closes is a valid/consistent behavior.

A few small changes are also made in order to improve debugging.