feat: add control over autonat (#176)
Prior to this change autonat would always be enabled. Now it can be
disabled in cases where NAT is not being used.

Additionally previously an unbounded number of probes would be running
concurrently for all observed addresses. When an dynamic outbound NAT is
used this results in many probes. Now we limit to one active probe at a
time and also track failed external address so we do not repeatedly
attempt to probe them.