fix: do not use bootstrap list with kademlia
Prior to the peer_manager connecting to bootstrap peers we relied on
kademlia to connect to them. Now not only is it not necessary but the
peer_manager applies an appropriate backoff for redials where kademlia
does not. There should be no change in external behavior except that we
no longer spam dials to disconnected bootstrap peers.

Now the peer_manager (renamed to ceramic_peer_manager) can report of a
peer id is a ceramic peer. If we connect to a ceramic peer we add it to
the kademlia routing table regardless of its reported protocols. This is
because we discovered that peers sometimes do not report they support
the kademlia protocol.