feat: add value support to Recon (#217)
With this change Recon is a key value store instead of simply a key
store.

Recon synchronization is the same in principle but its underlying
protocol has been refactored.

The protocol design is now has clear initiator and responder roles.
Generally the initiator makes requests and the responder responds.
Messages themselves are small and represent a small amount of work. This
way both nodes can be actively working concurrently to synchronize.
Prior to this change nodes would be idle while the other was working and
would frequently deadlock if a single message size grew too large.

Values are synchronized by sending ValueRequests and ValueResponses
along side synchronization messages.

The API into the Recon protocol is now a Stream + Sink allowing any
system that can transport messages in full duplex a possible transport
for Recon. The protocol implementation is no longer specific to libp2p.
This means we can use HTTP in the future. However this PR removes the
HTTP implementation as its not trivial with the HTTP server we are using
to create a full duplex channel.