chore: adding ci for cargo machete (#252)
* chore: adding ci for cargo machete

This adds a Makefile target to use machete to check for any unused
dependencies.

Its possible that machete has a false positive. For those cases simply
add the dependencies to the ignored list for machete in the Cargo.toml

In order to make machete happy we needed to clean up the generated
Cargo.toml files for the api-server and kubo-rpc-server crates.
Since these updates to the toml file are now more involved we use a
augtool to safely edit these files.

Builders now must install augeas/augtool.

* fix: do docker build natively