fix: refactor recon storage and remove sqlx dependency from recon and core crates (#254)
* fix: remove sqlx dependency from recon and core crates

Modified the store implementations to be interest/model specific as they have different requirements.
Interests don't have values, so those requests no-op (they error if you specify a value other than []).
Removed the sort_key column from the tables as its purpose was interest/model differentiation.
Renamed the tables accordingly and adjusted the queries.

* fix: remove sort_key from model_block table as well

* fix: Make model_block PK (cid,key)

it's (probably) possible to have keys point to the same CID, so make sure they are different rows

* fix: interest range_with_values query was not updated

In current feed PR, the API was modified to use range instead of range_with_values in this case, but it needed to be fixed here too

* chore: review clean up

* fix: interest store always has a value for a key, not never

---------

Co-authored-by: David Estes <dav1do@users.noreply.github.com>