fix: use try_from on recon keys (#263)
* fix: use try_from on recon keys

This changes the recon::Key trait to use TryFrom<Vec<u8>> instead of
simply From<_> in order to allow for validation of the bytes. Now its
_should_ be impossible to construct either an Interest or an EventId
that is not valid. Remembering that a fencepost is considered valid.

* fix: fix bug in decoding event id

The initial change had a bug in how the event id was decoded as it
required an event height which was not always present in fenceposts