feat: report error if event network does not match
Prior to this change events written via the API that do not match the
configured network of the node were stored but never synchronized as
interests are specific to the configured network.

Now an error is returned from the API if the network of the event does
not match the configured network.

NOTE: This will eventually be unneeded logic once event ids are removed
from the API.