fix: remove loud warn from Recon (#287)
Prior to this change we would emit a warning every time we connected to
a peer that did not speak the recon protocol. As we connect to many
peers for the DHT that do not speak Recon the warn was triggered a lot.
We do not need the warn as we have a metric counter for it and its not
generally helpful debugging information.