feat: add postgres tests
Some ugly macro use (hack-ro?) to run a test for all databases. The postgres tests only run if `PG_TESTS=1`. The intention being that we don't want `cargo test` to fail generally, but in CI we want to start a db and make sure things pass

I also updated the bitswap Store trait to include put, as we were going to the blockstore directly before