refactor: add service crate (#354)
* refactor: create new ceramic event store to encapsulate event parsing logic

intention is to remove trait impls from store crate in favor of service, and better isolate some of the store files

* refactor: split event/block sql logic into multiple files

* refactor: only impl store traits in the service crate, refactor store crate

The store crate now has entities to hold all the deserialization/table definition related code, and the access module provides a higher level API that the service crate can rely on to interact with the database.

* refactor: move event file to access folder

left for last commit to make it possible to see changes in git

* chore: remove comments

* refactor: adjust carfile rebuilding for recon vs api context

split types to ensure event ID and CID aren't getting mixed up right now. build events that look mostly valid for tests.moved some tests where we want to hard code CIDs to store so it doesn't care about event validation/structure

* chore: clean up

* refactor: hide sqlx details in store crate public api

* fix: fix missing else branch causing panics post shutdown

just an annoyance in the logs but we can do better

* fix: better fix for server insert task shutdown