fix: events incorrectly left undelivered (#366)
* fix: bug marking event deliverable

An event might be left hanging undelivered until an additional event came in for the stream. Now we check in memory events for a stream any time something is added to it.

* chore: docs

* fix: simplify select! logic

this will add more database reads for events on the rx_new channel, but it's easier to understand and less error prone

* fix: avoid dropping ready events during task cancellation

log pending stream count when we're not receiving new events