fix: recurse fully to rebuild entire undelivered stream history
it was possible to only deliver segments of the history up to a new tip, when we could keep going and find mulitple new tips in one pass. this meant could have needed a trigger after the last event to fully deliver everything. now we build the entire history if it's on the node regardless of the state of the delivered/undelivered segments.

this also uses an interval instead of an if in the loop to ensure it's triggered when nothing is incoming.