fix: IOD long streams could remain undelivered (#387)
* fix: correct/better IOD

- is now correct and handles long streams out of order successfully in all cases (afaict). Should be more performant as it only does work when it should be necessary.
- includes test of long streams

* fix: change ordering events function

- use map of {cid: deliverable} to reduce the number of entires we have to review at the end
- use a VecDeque instead of a Vec to fix a bug where it would just process the same event and exit rather than find the in memory history
- need to write tests to exercise this. it would have rejected an api write that had a prev in one batch while recon events would be sorted out. would be hard to encounter but merits tests, particulary because it's hard to encounter.

* refactor: rename structs with Meta instead of Header

* fix: fixes and changes based on review comments

- process all undelivered events at server startup so we can fail the process not just the task
- adjust loop to while condition
- update some comments/function names
- fail loudly (i.e. panic) on anything expected to be unreachable (next up: refactor to make those states unrepresentable)

* chore: add test logging

* refactor: modify IOD task to hopefully be more readable

- modified the approach to keep everything in the maps until we're done and change things in place when it's identified as deliverable.
this has the advantage of better cancel safety and more consistency about what is in the maps.
- also renamed some things from header->metadata

* fix: clean up docs/comments

* refactor: use event crate carfile parsing

also changed struct sent to ordering task to avoid sending body we don't need and the ordering task correspondingly

* fix: optimization to do less work

the check should have been cid_map before too

* fix: fix initial (shallow) ordering code and write tests

* fix: only panic on internal states in release mode

we should never discover undelivered init events while processing the start up batch, but we can correct them so we aren't going to crash

* chore: improve logging for start up processing of undelivered events

Don't want the process to appear to be hanging without any indication of what is going on if there is a large backlog of undelivered events

* chore: address PR review feedback

use `send` instead of `try_send` to enforce backpressure, clean up comments/docs

* fix: prune data to avoid unbounded growth of delivered events

if we're not done, we keep one event in memory until next time to optimize for in order streams. we also try to process on all undelivered events for a stream we're tracking as things fork and we don't know what happened while we were in the queue

* Update ordering_task.rs

Co-authored-by: Spencer T Brody <spencer+github@3box.io>

---------

Co-authored-by: Spencer T Brody <spencer+github@3box.io>