feat: change default dir to current directory (#431)
BREAKING CHANGE: Both the store and p2p key dirs now default to the
current working directory instead of the home/.ceramic-one directory.

Generally this makes it less error prone when running multiple daemon on
a single host as the naturally isolate from each other using their
working directory. Running multiple daemons locally is a common task in
order to experiment with the p2p nature of ceramic-one.

This is a breaking change. If you previously relied on the home
directory default you must now explicitly pass in the home directory
like so:

    ceramic-one daemon -s ~/.ceramic-one -p ~/.ceramic-one