chore: change network names
we now return the names used previously by js-ceramic rather than something that looks like a pubsub topic. this could be subtly breaking, but afaict, nothing uses or relies on this currently.

we could instead impl Display or another function to get this value, but given the unused nature I'm preferring one version unless others deem that a mistake.