fix: use default (google) dns servers if system resolv.conf is invalid
If the system dns conf is missing nameservers, dns queries will never be resolved. Currently, hickory-dns doesn't error in this case so we handle it manually until the fix is released. This happens on osx if wifi is disabled when you start c1 and is never able to resolve itself without this change.