feat: add noop ceramic-olap binary
The ceramic-olap binary copies verbatim some setup code from ceramic-one
re tracing/logging/metrics. Eventually this may move back into the
ceramic-one binary or we can find a way to make it generic and reuse.
Given this is only the first copy of the code I don't think we should
work to make it generic yet. Not until we either decide this is the long
term home for the ceramic-olap binary or have a third binary we want to
have a similar setup.

This binary does nothing yet except start tracing/logging/metrics and
then wait for a signal to shutdown gracefully.