refactor: remove recon client/server
We no longer need exclusive access to recon so removing the client/server loop abstraction as it currently acts as an extra layer to update. I added Clone to a couple things and had to adjust generics.. I think it's okay but probably merits a close review.

The tests now use worker threads instead of spawning the server in the background. This changed the order that things get sent, but it looks like all the same messages are exchanged.