feat: remove persisting event batch when looking at ranges
this is a modification to the previous commit that actually batches writes until the conversation finishes, however, it changes the semantic a bit as we no longer send the entire hash range. We could probably recognize that range and persist to make sure we're in sync at the end, but if we're okay with this, we won't do that.

It does make the tests look a bit odd as we get values but our state isn't updated until the conversation finishes and we're in sync. I can change the batch size to 1 if we want this but would prefer to see the peer state updated in the expect output