refactor: remove recon client/server and send empty ranges when we don't need anything (#509)
* refactor: remove recon client/server

We no longer need exclusive access to recon so removing the client/server loop abstraction as it currently acts as an extra layer to update. I added Clone to a couple things and had to adjust generics.. I think it's okay but probably merits a close review.

The tests now use worker threads instead of spawning the server in the background. This changed the order that things get sent, but it looks like all the same messages are exchanged.

* feat: reply with only necessary ranges to discover bounding keys when the remote was missing a range

this appears to have reduced round trips in tests without changing the synchronization result

* feat: remove persisting event batch when looking at ranges

this is a modification to the previous commit that actually batches writes until the conversation finishes, however, it changes the semantic a bit as we no longer send the entire hash range. We could probably recognize that range and persist to make sure we're in sync at the end, but if we're okay with this, we won't do that.

It does make the tests look a bit odd as we get values but our state isn't updated until the conversation finishes and we're in sync. I can change the batch size to 1 if we want this but would prefer to see the peer state updated in the expect output

* refactor: remove the need for reply_with (#510)

* chore: clean up comment and remove uneeded clone

---------

Co-authored-by: Nathaniel Cook <nvcook42@gmail.com>