refactor: create explicit event and interest services with separate stores. (#513)
* refactor: move ceramic-store into ceramic-service

The ceramic-store crate was only consumed by the ceramic-service. In
order to establish a clear pattern that each service should have its own
access path into sqlite we move the ceramic-store crate to be a
submodule of ceramic-service.

* refactor: split ceramic-service into event-svc and interest-svc

* refactor: add ceramic_sql crate for shared sqlite pool

All services will need access to sqlite. This means they need to share a
sqlite pool type. This pulls out the sqlite pool from the services into
a shared type.

* refactor: use store and service names consistently

A store talks directly to a database. A service provides a public API
and consumes a store.

* refactor: fix bugs in rename

* refactor: fix clippy

* refactor: remove unused metrics middleware