refactor: adjust event store to return pointer to inserted event and adjust internal service insert result
this reduces the data we copy and reduces the amount of comparisions and lookups we have to do to try to find
the information we need to ordering (was it new, what event was this?). now we just get the pointer back so
it's easy to know which event was new or not and find any information about it we care about.