feat: add webauthn signature validation (#511)
This is breaking changes to the cacao/capability types, but those should only be internal for now.

The encoding of the metadata is pretty awful, not sure if I'm just missing a nice way to do it. Also don't really like adding an enum for multiple signature metadata types, but these don't include any key/alg, so we could add IPLD as a MetadataValue, but would have to make the key/alg optional or have a default.