refactor: return reference to inserted item from event store and reduce unnecessary ordering work  (#530)
* refactor: adjust conclusion test to use api instead of recon store (stronger ordering requirements)

* refactor: set insertable as deliverable for init events

moved the parameter earlier so we can borrow from the event before it's moved without having to assign a temp variable

* refactor: split order events function into ordering with db and just in memory

now we do fewer lookups for recon items at insert time and offload to the ordering task to deal with

* feat: skip all ordering (in memory and db) for lazy deliverable (migrations)

* refactor: adjust event store to return pointer to inserted event and adjust internal service insert result

this reduces the data we copy and reduces the amount of comparisions and lookups we have to do to try to find
the information we need to ordering (was it new, what event was this?). now we just get the pointer back so
it's easy to know which event was new or not and find any information about it we care about.

* refactor: remove unnecessary field from store insert result

* refactor: add back removed Eq derive