feat: add example hoku time event validation
This is mostly an example and requires figuring out the chain ID information and time event format, as well as passing in URLs via CLI flags, but demonstrates how the client could be hooked up using the refactored trait. It would also require some modifications to the exports in the hoku sdk to work (e.g. Address, HttpClient, FvmQueryHeight).