feat: teach query command how to read the tables
There are currently two tables in the datafusion _pipeline_
architecture, conclucsion_feed and doc_state.

This change moves the initialization code for datafusion to interact
with those tables into a new ceramic-pipeline crate where both the olap
and query commands can reuse the logic. This new crate can be a home for
future work in this _pipeline_ architecture as we flesh it out. For
example we will likely move the aggregator/ceramic_patch logic into this crate.

In either case the query command now knows how to query both tables when
it first starts up. This table access is via flight sql and direct
object store access.