fix: use correct delivered index
The previous code would zip the parsed events with the all_blocks
iterator. However the all_blocks iterator was longer than parsed and so
the delivered values got misaligned with the events, causing the index
of conclusion feed events to be incorrect.

Now we correctly filter down the all_block iterator so it has only a
single value per event.