refactor: make event access stateful
Previously a global static delivered counter was used to assign/track
delivered values. However this makes testing hard as if multiple tests
run they effect the delivered values of each other.

This change makes the EventAccess contain the counter state within
itself and then anywhere that a pool was used to access events it was
replaced with an Arc<EventAccess>. This means we can truly have
isolated delivered counters.