feat: use data container wrapper for mutable metadata (#559)
* feat: use data container wrapper for mutable metadata

Existing MID events have a header on data payloads. This data is mutable
metadata. We needed a mechanism to represent this kind of data within
the conclusion events and doc states. We added a simple wrapper type
with metadata and data fields.

* fix: update tests after rebase

* fix: treat metadata as transitive

Instead of expecting each data event header to repeat the metadata we
carry forward the metadata from the previous state and only if its
present do we apply a change to the metadata. Tests have been updated to
reflect this new behavior.

* update test

* refactor: small code reorg for clarity