feat: backoff recon synchronization to peers with bad data
With this change peers will backoff synchronization attempts with peers
that fail synchronization. This will have the effect that the bootstrap
nodes do not talk to nodes that are sending bad data very often.

However the backoff state is tied to the connection state and is reset
if the connection resets. Allowing developers to restart their node and
be in good standing with bootstrap nodes.

Additionally a small change was added to prefer using the connected
address to a peer when reporting the address of connected peers.