fix: keep interests and model syncs separate
This change makes it so that interests and model syncs are independent.
We always start with the interests sync. But then after the initial
syncs they can independently be backed off.

This helps with no resetting idle connections and therefore resetting the
backoff state. As there is likely some activity on the connection for
one of the two syncs.