fix: handle double write of same data (#613)
Prior to this change a write of the same data (keyed by the event id)
would cause an error the batching logic of the write path. It was
assumed that double writes would not happen and so the channel to report
the insert result was dropped, causing an error.

Now all channels are preserved and the result is sent to all.