feat: adds /peers endpoint
With this change a new API endpoint is added for accessing connected
peers. The API is similar to the /api/v0/swarm/peers endpoint however it
allows for multiple addresses and does not have to remain Kubo RPC
compatible.

The endpoint allows posting an address and therefore connecting to a new
peer.

Fixes: #608