feat: stop synchronizing interests
With this change the interests ring is no longer synchronized between
peers. A local node will still use the interest-svc to know its own
interests and the beginning of each Recon conversation negotiates shared
interests. Therefore it is no longer necessary to synchronize interests.

In the future we may decide that we want to use Recon to sync interests
instead of a linear sharing of interests before each conversation,
however that is only a performance optimization that is not important at
this stage.

Fixes #610
Fixes #611