chore: use explicit os versions
This week github updated the `ubuntu-latest` tag to be ubuntu-24.04.
This meant that version 0.47.0 of ceramic-one no longer runs on
ubuntu-22. This is unexpected and not in our control. With this change
we explicitly name the os versions we target (this matters because of
glibc versions). This means we will need to add new versions and
deprecate old versions as they are released.