fix: fix off by one, projection and not found bugs (#647)
There was an off by one bug in translating highwater marks to offsets.

There was a bug in projecting stream tables as they reported the wrong
schema.

Finally there was a bug in the stream state query that would panic if
the stream didn't exist and datafusion returned an empty batch.