refactor: more efficient list udf impls
This is a more efficient list UDF implementation when the list structure
does not change. We simply map over the values array directly and reuse
the list offsets and nulls.

This is not strictly necessary. However setting a standard to always
operate on a single Arrow array at a time is kinda the whole point of
Arrow. Therefore I thinks is valuable to make this change so we set a
good pattern to follow in the future.