refactor: use explicit order columns
Prior to this change it was assumed that the index column was a global
ordering value for all tables in the pipeline. However that cannot be
true as event_states have stricter order constraints than conclusion
events. As such this change breaks that pattern with explicit order
columns and their meaning.

This is prework to be able to correctly buffer and reorder event states
according to cross stream dependencies.