fix: remove distinct step as its prohibitively expensive (#682)
The distinct step when fetching pending events was to correct for the
case when we failed to delete old pending events. However it was
prohibitively expensive consuming many GBs of memory for roughly 200K
pending event rows.

There are few reasons why the distinct was too aggregisve and no the
best solution to the duplicate problem.

1. While duplicate pending events can arise seeing the same model cid
   twice is not possible. This means that in worst case scenario pending
   events get processed once but fail to be deleted, in which case they
   remain in the pending events table forever. But those duplicates will
   not result in event state rows
2. With retries (now added) the likely hood of duplicates pending events
   is very small and not worth paying the de-duplication cost on every
   read of the table.

With this change we retry any failures to delete old pending events and
explicitly log any errors. We no longer de-duplicate pending events
on read. In a worst case scenario an operator can delete the event_states
and pending_event_states tables and rebuild all the data to remove the
pending events.

Fixes: #680