fix: update ordering and push down filters
we now pass predicates down into the pipeline so that _feed queries work better.
if conclusion_event_order is included in the predicate, we will skip reading all events
and start from the offset (we could potentially push the expr down but it's more involved).
in general, WHERE clause predicates should be respected by _feed queries now. I had to adjust
the ordering used as we were conflating conclusion_event_order and event_state_order, so we now
use the conclusion_event_order as the offset, and apply predicates afterward to filter the stream
data returned. afaict, the adjustment to use >= instead of > is required now, and doesn't cause
issues as we expect we may get the same event in both streams and reduce to one correctly.