fix: random things like feed queries, query CLI, migration logging, deps (#708)
* chore: update flight sql deps

* fix: check migration model filter earlier

this avoids lots of errors e.g. "fatal error: finding capability block" for events we don't want to review

* fix: adjust query CLI to display streaming results

querying the _feed tables without a limit will now print all the data retrieved and then wait for updates. previously it would just hang doing nothing.

* test: modify tests so that events pass validation

* fix: index -> conclusion_event_order

this column was renamed but the predicate was not correctly updated so the feed queries always started from the beginning

* fix: revert predicate pushdown change

* fix: update ordering and push down filters

we now pass predicates down into the pipeline so that _feed queries work better.
if conclusion_event_order is included in the predicate, we will skip reading all events
and start from the offset (we could potentially push the expr down but it's more involved).
in general, WHERE clause predicates should be respected by _feed queries now. I had to adjust
the ordering used as we were conflating conclusion_event_order and event_state_order, so we now
use the conclusion_event_order as the offset, and apply predicates afterward to filter the stream
data returned. afaict, the adjustment to use >= instead of > is required now, and doesn't cause
issues as we expect we may get the same event in both streams and reduce to one correctly.

* chore: point to right sdk docs

* fix: improve stream/feed queries to support filters

previously, we only supported highwater mark and tried to filter after
the fact. now, we create two streams, the first using the filters to
find all existing events, the second is all new events that are filtered
before being pushed to the consumers.

* chore: Option vec instead of allocating empty vec for filters

* chore: fix clippy lints

* fix: return error instead of simply logging when expression can't be evaluated

* chore: update sccache version

* fix: use SessionContext in feed to use execution props

* chore: clippy

* fix: remove TODO

---------

Co-authored-by: Mohsin Zaidi <mohsinrzaidi@gmail.com>